Refactor checkers to allow more Pydantic approach later
